+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.33] - 2026-04-09
+
+### Added
+- Code block copy button with localization support (Copy/复制, Copied!/已复制!)
+- Link URL tooltip on hover
+- Image lightbox on click
+- Heading anchor links
+
 ## [1.0.32] - 2026-04-09
 
 ### Fixed